+++ src/tests/ftest/container/attribute.py
@@ -32,7 +32,7 @@ from apricot import TestWithServers
 
 from general_utils import DaosTestError
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 # pylint: disable = global-variable-not-assigned, global-statement
 GLOB_SIGNAL = None
