+++ src/tests/ftest/container/basic_snapshot.py
@@ -39,8 +39,8 @@ sys.path.append('./../../utils/py')
 # pylint: disable=import-error
 import server_utils
 import write_host_file
-from daos_api import (DaosContext, DaosPool, DaosContainer, DaosSnapshot,
-                      DaosLog, DaosApiError)
+from pydaos.raw import (DaosContext, DaosPool, DaosContainer, DaosSnapshot,
+                        DaosLog, DaosApiError)
 
 # pylint: disable=broad-except
 class BasicSnapshot(Test):
