+++ src/tests/ftest/container/create.py
@@ -29,7 +29,7 @@ import uuid
 
 from apricot import TestWithServers
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 
 class CreateContainerTest(TestWithServers):
