+++ src/tests/ftest/container/open.py
@@ -30,7 +30,7 @@ import uuid
 from avocado import main
 from apricot import TestWithServers
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 
 class OpenContainerTest(TestWithServers):
