+++ src/tests/ftest/container/open_close.py
@@ -28,7 +28,7 @@ import traceback
 import uuid
 from apricot import TestWithServers
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 
 class OpenClose(TestWithServers):
