+++ src/tests/ftest/object/array_obj_test.py
@@ -31,8 +31,7 @@ from avocado import main
 from apricot import TestWithServers
 
 
-from conversion import c_uuid_to_str
-from daos_api import DaosPool, DaosContainer, DaosApiError
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError, c_uuid_to_str
 
 class ArrayObjTest(TestWithServers):
     """
