+++ src/tests/ftest/object/obj_open_bad_param.py
@@ -27,9 +27,7 @@ import os
 import traceback
 from apricot import TestWithServers
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
-from daos_cref import DaosObjId
-
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError, DaosObjId
 
 class ObjOpenBadParam(TestWithServers):
     """
