+++ src/tests/ftest/object/obj_update_bad_param.py
@@ -30,7 +30,7 @@ from avocado import main
 from apricot import TestWithServers
 
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 class ObjUpdateBadParam(TestWithServers):
     """
