+++ src/tests/ftest/object/punch_test.py
@@ -27,8 +27,7 @@ import os
 import traceback
 
 from apricot import TestWithServers
-from daos_api import DaosPool, DaosContainer, DaosApiError
-
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 class PunchTest(TestWithServers):
     """
