+++ src/tests/ftest/object/same_key_different_value.py
@@ -28,8 +28,7 @@ import time
 import traceback
 from apricot import TestWithServers
 
-from daos_api import DaosPool, DaosContainer, DaosApiError
-
+from pydaos.raw import DaosPool, DaosContainer, DaosApiError
 
 class SameKeyDifferentValue(TestWithServers):
     """
