+++ src/tests/ftest/pool/permission.py
@@ -31,7 +31,7 @@ from apricot import Test
 import agent_utils
 import server_utils
 import write_host_file
-from daos_api import DaosContext, DaosPool, DaosContainer, DaosLog, DaosApiError
+from pydaos.raw import DaosContext, DaosPool, DaosContainer, DaosLog, DaosApiError
 
 class Permission(Test):
     """
