+++ src/tests/ftest/util/apricot/apricot/test.py
@@ -39,7 +39,7 @@ import fault_config_utils
 import agent_utils
 import server_utils
 import write_host_file
-from daos_api import DaosContext, DaosLog, DaosApiError
+from pydaos.raw import DaosContext, DaosLog, DaosApiError
 
 
 # pylint: disable=invalid-name
